# 1. Distribución/API/APIBANCOV1/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "APIBANCOV1, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/banco/api/v1/Autenticacion/validacion": {
      "post": {
        "tags": [
          "Autenticacion"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AuthBE"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/banco/api/v1/Cuenta": {
      "get": {
        "tags": [
          "Cuenta"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Cuenta"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CuentaBE"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/banco/api/v1/Cuenta/{id}": {
      "get": {
        "tags": [
          "Cuenta"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Cuenta"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CuentaBE"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Cuenta"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/banco/api/v1/Movimiento": {
      "get": {
        "tags": [
          "Movimiento"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Movimiento"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/MovimientoBE"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/banco/api/v1/Movimiento/{id}": {
      "get": {
        "tags": [
          "Movimiento"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Movimiento"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/MovimientoBE"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Movimiento"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/banco/api/v1/Tarjeta": {
      "get": {
        "tags": [
          "Tarjeta"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Tarjeta"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TarjetaBE"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/banco/api/v1/Tarjeta/{id}": {
      "get": {
        "tags": [
          "Tarjeta"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Tarjeta"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TarjetaBE"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Tarjeta"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/banco/api/v1/Usuario": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UsuarioBE"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/banco/api/v1/Usuario/{id}": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Usuario"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UsuarioBE"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Usuario"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "AuthBE": {
      "type": "object",
      "properties": {
        "usuario": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "CuentaBE": {
      "type": "object",
      "properties": {
        "usuarioId": {
          "format": "uuid",
          "type": "string"
        },
        "numeroCuenta": {
          "type": "string"
        },
        "monedaCuenta": {
          "$ref": "#/definitions/Moneda"
        },
        "tipoCuenta": {
          "$ref": "#/definitions/Tipo"
        },
        "linea": {
          "format": "double",
          "type": "number"
        },
        "saldo": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "Moneda": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3
      ],
      "type": "integer"
    },
    "MovimientoBE": {
      "type": "object",
      "properties": {
        "cuentaId": {
          "format": "uuid",
          "type": "string"
        },
        "monto": {
          "format": "double",
          "type": "number"
        },
        "origen": {
          "$ref": "#/definitions/Origen"
        }
      }
    },
    "Origen": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3,
        4
      ],
      "type": "integer"
    },
    "TarjetaBE": {
      "type": "object",
      "properties": {
        "cuentaId": {
          "format": "uuid",
          "type": "string"
        },
        "usuarioId": {
          "format": "uuid",
          "type": "string"
        },
        "numeroTarjeta": {
          "type": "string"
        },
        "mes": {
          "format": "int32",
          "type": "integer"
        },
        "año": {
          "format": "int32",
          "type": "integer"
        },
        "cvv": {
          "type": "string"
        },
        "pin": {
          "type": "string"
        }
      }
    },
    "Tipo": {
      "format": "int32",
      "enum": [
        0,
        1
      ],
      "type": "integer"
    },
    "UsuarioBE": {
      "type": "object",
      "properties": {
        "nombres": {
          "type": "string"
        },
        "apellidos": {
          "type": "string"
        },
        "edad": {
          "type": "string"
        },
        "contraseñaWeb": {
          "type": "string"
        }
      }
    }
  }
}