# 2. Negocio/BancoEntity/UsuarioBE.cs
namespace BancoEntity
{
    public class UsuarioBE
    {
        public string Nombres { get; set; }
        public string Apellidos { get; set; }
        public string Edad { get; set; }
        public string ContraseñaWeb { get; set; }
    }
}